+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# consumers-2.2.0.0 (2019-02-19)
+* When jobs are relased from a dead consumer, reset their finished_at column
+
 # consumers-2.1.2.0 (2018-07-11)
 
 * Support hpqtypes-1.6.0.0.